Portugal municipal census data 2011

The 2011 Portugal Census data contains the centroid coordinates and 12 selected socio-economic variables for the 278 municipalities of Portugal.The 12 selected variables are denspop11(Population density) discapd(Average distance to the closest district capital) gmm2011(Average monthly earnings) podc(Daily purchasing power per capita) tdesmp11(Unemployment rate) pbsdt11(Population receiving the unemployment benefit) ne_sup(Population with a university degree completed) abanesc(Percentage of school dropout) pdmcdt11(Expenses of the municipality on cultural and sports activities) pdiv(Percentage of divorces) distpolall(Average distance to the closest police station) tcrmiltot(Criminality rate)

Usage

pmc

Format

pmc: A tibble of 2011 Portugal Census data with 278 rows and 14 variables.